Move world-data systems into unified folder structure
Group datatiles/, wind/, water/, and terrain/ into src/game/world-data/
to reflect their shared architectural patterns:
- Entity-based *Info classes with identical lifecycle patterns
- Querier interfaces for collecting query forecasts
- Hybrid GPU/CPU computation with identical folder structure
- DataTileComputePipeline integration
