+++ src/game/boat/Keel.ts
@@ -8,7 +8,7 @@ import {
   foilLift,
   KEEL_CHORD,
 } from "../fluid-dynamics";
-import { WaterInfo } from "../water/WaterInfo";
+import { WaterInfo } from "../world-data/water/WaterInfo";
 import { KeelConfig } from "./BoatConfig";
 import { Hull } from "./Hull";
