+++ src/game/boat/Rudder.ts
@@ -8,7 +8,7 @@ import {
   foilLift,
   RUDDER_CHORD,
 } from "../fluid-dynamics";
-import { WaterInfo } from "../water/WaterInfo";
+import { WaterInfo } from "../world-data/water/WaterInfo";
 import { RudderConfig } from "./BoatConfig";
 import { Hull } from "./Hull";
