+++ src/game/world-data/terrain/TerrainQuerier.ts
