+++ src/game/world-data/water/WaterInfo.ts
@@ -8,12 +8,12 @@
  */
 
 import { createNoise3D, NoiseFunction3D } from "simplex-noise";
-import BaseEntity from "../../core/entity/BaseEntity";
-import { on } from "../../core/entity/handler";
-import Game from "../../core/Game";
-import { profile } from "../../core/util/Profiler";
-import { SparseSpatialHash } from "../../core/util/SparseSpatialHash";
-import { V, V2d } from "../../core/Vector";
+import BaseEntity from "../../../core/entity/BaseEntity";
+import { on } from "../../../core/entity/handler";
+import Game from "../../../core/Game";
+import { profile } from "../../../core/util/Profiler";
+import { SparseSpatialHash } from "../../../core/util/SparseSpatialHash";
+import { V, V2d } from "../../../core/Vector";
 import {
   DataTileComputePipeline,
   DataTilePipelineConfig,
