+++ src/game/world-data/wind/WindConstants.ts
