+++ src/game/world-data/wind/WindInfo.ts
@@ -10,12 +10,12 @@
  */
 
 import { createNoise3D, NoiseFunction3D } from "simplex-noise";
-import BaseEntity from "../../core/entity/BaseEntity";
-import { on } from "../../core/entity/handler";
-import Game from "../../core/Game";
-import { profile } from "../../core/util/Profiler";
-import { SparseSpatialHash } from "../../core/util/SparseSpatialHash";
-import { V, V2d } from "../../core/Vector";
+import BaseEntity from "../../../core/entity/BaseEntity";
+import { on } from "../../../core/entity/handler";
+import Game from "../../../core/Game";
+import { profile } from "../../../core/util/Profiler";
+import { SparseSpatialHash } from "../../../core/util/SparseSpatialHash";
+import { V, V2d } from "../../../core/Vector";
 import {
   DataTileComputePipeline,
   DataTilePipelineConfig,
@@ -26,7 +26,7 @@ import type {
   QueryForecast,
   ReadbackViewport,
 } from "../datatiles/DataTileTypes";
-import { isWindModifier, type WindModifier } from "../WindModifier";
+import { isWindModifier, type WindModifier } from "../../WindModifier";
 import {
   computeBaseWindAtPoint,
   WindComputeParams,
