+++ src/game/world-data/wind/WindQuerier.ts
